rename and migrate testmodel
