+++ tests/test_statusfield.py
@@ -5,6 +5,7 @@ from django.forms import ChoiceField
 
 from dkmodelfields.statusfield import StatusField, StatusValue
 from django.utils.translation import ugettext_lazy as _
+from testapp_dkmodelfields.models import S
 
 
 def test_status_field():
@@ -72,3 +73,8 @@ def test_status_field():
     )
 
     assert isinstance(sf.formfield(), ChoiceField)
+
+
+def test_s_model(db):
+    s = S.objects.create()
+    assert type(s.status) == str
