+++ tests/testapp_dkmodelfields/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TestAppConfig(AppConfig):
+    name = 'testapp_dkmodelfields'
+    verbose_name = 'TestApp dkmodelfields'
